requirements.txt
requests==2.31.0
pytest==7.4.3
pydantic==2.5.2
python-dotenv==1.0.0
assertpy==1.1
pytest-html==4.1.1
allure-pytest==2.13.2
pytest-order==1.1.0
typing-extensions==4.8.0
